Seperate app with listen code for cleaner code testing
